hiring_assistant/requirements.txt
streamlit
groq>=0.5.0
python-dotenv
